UI調整: タイトルとガイドパネルの更新
