+++ README.md
@@ -1,4 +1,4 @@
-# ボルダリングポーズ検討アプリ
+# ルートプランナー
 
 実在するボルダリングコース（写真）を背景に、Magic Poserのような操作性で人物モデルのポーズを段階的に試すことで攻略手順を検討するPCブラウザ向けアプリです。
